requirements.txt
opencv-python
mediapipe==0.10.32
scikit-learn
numpy
Pillow
